Consolidate all calls under the client
