+++ src/cli/cmd/mod.rs
@@ -8,8 +8,12 @@ pub(crate) mod login;
 pub(crate) mod search;
 pub(crate) mod status;
 
+use color_eyre::eyre::WrapErr;
 use once_cell::sync::Lazy;
-use reqwest::header::{HeaderMap, HeaderValue, ACCEPT, AUTHORIZATION};
+use reqwest::{
+    header::{HeaderMap, HeaderValue, ACCEPT, AUTHORIZATION},
+    Client,
+};
 use serde::{Deserialize, Serialize};
 use tabled::settings::{
     style::{HorizontalLine, On, VerticalLineIter},
@@ -107,6 +111,20 @@ pub(crate) enum FhError {
     Xdg(#[from] xdg::BaseDirectoriesError),
 }
 
+#[derive(Debug, Deserialize)]
+struct ProjectMetadata {
+    source_github_owner_repo_pair: String,
+    source_subdirectory: Option<String>,
+    version: String,
+}
+
+#[derive(Debug, Deserialize)]
+struct ProjectCanonicalNames {
+    project: String,
+    // FIXME: detect Nix version and strip .tar.gz if it supports it
+    pretty_download_url: url::Url,
+}
+
 impl FlakeHubClient {
     pub(crate) fn new(api_addr: &url::Url) -> Self {
         Self(api_addr.to_string())
@@ -117,24 +135,7 @@ impl FlakeHubClient {
         params: Option<Vec<(&str, String)>>,
         authenticated: bool,
     ) -> Result<T, FhError> {
-        let mut headers = HeaderMap::new();
-        headers.insert(ACCEPT, HeaderValue::from_static("application/json"));
-
-        if authenticated {
-            if let Ok(token) = tokio::fs::read_to_string(auth_token_path()?).await {
-                if !token.is_empty() {
-                    headers.insert(
-                        AUTHORIZATION,
-                        HeaderValue::from_str(&format!("Bearer {}", token.trim()))?,
-                    );
-                }
-            }
-        }
-
-        let client = reqwest::Client::builder()
-            .user_agent(crate::APP_USER_AGENT)
-            .default_headers(headers)
-            .build()?;
+        let client = make_base_client(authenticated).await?;
 
         let req = client.get(url);
 
@@ -183,12 +184,74 @@ impl FlakeHubClient {
 
         Self::get(url, None, true).await
     }
+
+    async fn metadata(
+        &self,
+        org: &str,
+        project: &str,
+        version: &str,
+    ) -> color_eyre::Result<ProjectMetadata> {
+        let url = flakehub_url!(&self.0, "version", org, project, version);
+        let client = make_base_client(true).await?;
+
+        let res = client.get(&url.to_string()).send().await?;
+
+        if let Err(e) = res.error_for_status_ref() {
+            let err_text = res.text().await?;
+            return Err(e).wrap_err(err_text)?;
+        };
+
+        let res = res.json::<ProjectMetadata>().await?;
+
+        Ok(res)
+    }
+
+    async fn project_and_url(
+        &self,
+        org: &str,
+        project: &str,
+        version: Option<&str>,
+    ) -> color_eyre::Result<(String, url::Url)> {
+        let url = match version {
+            Some(version) => flakehub_url!(&self.0, "version", org, project, version),
+            None => flakehub_url!(&self.0, "f", org, project),
+        };
+        let client = make_base_client(true).await?;
+        let res = client.get(&url.to_string()).send().await?;
+        if let Err(e) = res.error_for_status_ref() {
+            let err_text = res.text().await?;
+            return Err(e).wrap_err(err_text)?;
+        };
+        let res = res.json::<ProjectCanonicalNames>().await?;
+        Ok((res.project, res.pretty_download_url))
+    }
 }
 
 pub(crate) fn print_json<T: Serialize>(value: T) -> Result<(), FhError> {
     let json = serde_json::to_string(&value)?;
     println!("{}", json);
     Ok(())
+}
+
+async fn make_base_client(authenticated: bool) -> Result<Client, FhError> {
+    let mut headers = HeaderMap::new();
+    headers.insert(ACCEPT, HeaderValue::from_static("application/json"));
+
+    if authenticated {
+        if let Ok(token) = tokio::fs::read_to_string(auth_token_path()?).await {
+            if !token.is_empty() {
+                headers.insert(
+                    AUTHORIZATION,
+                    HeaderValue::from_str(&format!("Bearer {}", token.trim()))?,
+                );
+            }
+        }
+    }
+
+    Ok(reqwest::Client::builder()
+        .user_agent(crate::APP_USER_AGENT)
+        .default_headers(headers)
+        .build()?)
 }
 
 #[macro_export]
